Add linear interpolation

+++ include/nifty/transformation/affine_transformation.hxx
@@ -8,7 +8,7 @@ namespace transformation {
 
     template<unsigned NDIM, class MATRIX>
     inline void affineCoordinateTransformation(const array::StaticArray<int64_t, NDIM> & inCoord,
-                                               array::StaticArray<float, NDIM> & coord,
+                                               array::StaticArray<double, NDIM> & coord,
                                                const MATRIX & matrix) {
         for(unsigned di = 0; di < NDIM; ++di) {
             coord[di] = 0.;
